% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animal_GEI_LU.R
\docType{package}
\name{AnimalGEILU-package}
\alias{AnimalGEILU}
\alias{AnimalGEILU-package}
\title{Animal_GEI_LU: Package for Animal Greenhouse Gas Emissions and Land Use}
\description{
This package provides tools for calculating and analyzing greenhouse gas
emissions and land use metrics for animal systems.
}
\details{
For more information, see:
\itemize{
\item GitHub: \url{https://github.com/JGCRI/Animal_GEI_LU}
\item Webpage: \url{https://jgcri.github.io/Animal_GEI_LU/}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JGCRI/AnimalGEILU}
  \item Report bugs at \url{https://github.com/JGCRI/AnimalGEILU/issues}
}

}
\author{
\strong{Maintainer}: Zarrar Khan \email{zarrar.khan@pnnl.gov} (\href{https://orcid.org/0000-0002-8147-8553}{ORCID})

}
\keyword{internal}
