% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_N2O_direct_manure.R
\name{calculate_N2O_direct_manure}
\alias{calculate_N2O_direct_manure}
\title{Calculate direct N2O emissions from manure}
\usage{
calculate_N2O_direct_manure(
  automatic_cycle = FALSE,
  saveoutput = TRUE,
  data_dir = "user_data"
)
}
\arguments{
\item{automatic_cycle}{Logical. If TRUE, uses the built-in model for automatic farm cycle calculation. Default is FALSE.}

\item{saveoutput}{If TRUE (default) the results are saved in the output folder.}

\item{data_dir}{Path to the directory containing input CSV/data files. Defaults to `"user_data"`.}
}
\description{
Computes direct N2O emissions based on nitrogen excretion logic,
emission factors, management system, and climate (IPCC Eq 10.25).
}
