% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_metabolizable_energy.R
\name{calculate_poultry_energy}
\alias{calculate_poultry_energy}
\title{Calculate Poultry Metabolic Energy Requirements (FEDNA)}
\usage{
calculate_poultry_energy(saveoutput = TRUE)
}
\arguments{
\item{saveoutput}{If TRUE (default) the results are saved in the output folder.}
}
\description{
Calculate Poultry Metabolic Energy Requirements (FEDNA)
}
