% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_VS_manure.R
\name{calculate_vs}
\alias{calculate_vs}
\title{Calculate Volatile Solids (VS) for Animals}
\usage{
calculate_vs(urinary_energy = 0.04, saveoutput = TRUE, data_dir = "user_data")
}
\arguments{
\item{urinary_energy}{Numeric. Fraction of energy lost in urine. Default 0.04.}

\item{saveoutput}{If TRUE (default) the results are saved in the output folder.}

\item{data_dir}{Character. Path to the folder containing input CSV files. Default is \code{"user_data"}.}
}
\value{
Tibble with VS for all animal categories.
}
\description{
Computes volatile solids (VS) based on Gross Energy (GE), Digestible Energy (DE),
and Ash content using IPCC Tier 2 methodology for ruminants, and DMI/Energy ratios for poultry.
}
