% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{categories}
\alias{categories}
\title{Datos de categorías}
\format{
Un data frame o lista con la clasificación de categorías.
}
\usage{
categories
}
\description{
Datos de categorías
}
\keyword{datasets}
