% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch4_mm}
\alias{ch4_mm}
\title{Mediciones de CH4 (Metano)}
\format{
Un conjunto de datos con registros de emisiones de metano.
}
\usage{
ch4_mm
}
\description{
Mediciones de CH4 (Metano)
}
\keyword{datasets}
