% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{characteristics}
\alias{characteristics}
\title{Características del sistema}
\format{
Un conjunto de datos con atributos específicos del estudio.
}
\usage{
characteristics
}
\description{
Características del sistema
}
\keyword{datasets}
