% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefficients}
\alias{coefficients}
\title{Coeficientes de cálculo}
\format{
Un vector o data frame con coeficientes técnicos.
}
\usage{
coefficients
}
\description{
Coeficientes de cálculo
}
\keyword{datasets}
