% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crops}
\alias{crops}
\title{Datos de cultivos (crops)}
\format{
Información sobre tipos de cultivos y sus propiedades.
}
\usage{
crops
}
\description{
Datos de cultivos (crops)
}
\keyword{datasets}
