% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_user_data.R
\name{download_templates}
\alias{download_templates}
\title{Download Template Files from GitHub}
\usage{
download_templates(dest_folder = "user_data")
}
\arguments{
\item{dest_folder}{The name of the folder where the templates will be saved.
Defaults to "user_data".}
}
\value{
The absolute path to the `dest_folder` where the files were saved.
}
\description{
Downloads example data files (templates) from the GitHub repository and
saves them to a local folder.
}
\examples{
\dontrun{
 # Call the function to download the files
 data_path <- download_templates()

 # View the downloaded files
 list.files(data_path)

 # Read one of the files
 read.csv(file.path(data_path, "categories.csv"))
}

}
