% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emission_factors_direct}
\alias{emission_factors_direct}
\title{Factores de emisión directa}
\format{
Valores utilizados para calcular emisiones directas.
}
\usage{
emission_factors_direct
}
\description{
Factores de emisión directa
}
\keyword{datasets}
