% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emission_factors_indirect}
\alias{emission_factors_indirect}
\title{Factores de emisión indirecta}
\format{
Valores utilizados para calcular emisiones indirectas.
}
\usage{
emission_factors_indirect
}
\description{
Factores de emisión indirecta
}
\keyword{datasets}
