% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fractions}
\alias{fractions}
\title{Fracciones de conversión}
\format{
Valores proporcionales o fracciones utilizadas en el modelo.
}
\usage{
fractions
}
\description{
Fracciones de conversión
}
\keyword{datasets}
