% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_imports.R
\name{herdr-imports}
\alias{herdr-imports}
\title{Internal package imports}
\description{
This file imports the necessary functions from external packages
so you can use them without :: in your package functions.
}
\keyword{internal}
