% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herdr-package.R
\docType{package}
\name{herdr-package}
\alias{herdr}
\alias{herdr-package}
\title{herdr: Package for Animal Greenhouse Gas Emissions and Land Use}
\description{
This package provides tools for calculating and analyzing greenhouse gas
emissions and land use metrics for animal systems.
}
\details{
For more information, see:
\itemize{
\item GitHub: \url{https://github.com/JuanCBM99/herdr}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://juancbm99.github.io/herdr}
}

}
\author{
\strong{Maintainer}: Juan Carlos Baez \email{juancarlos.baez@bc3research.org} (\href{https://orcid.org/0000-0002-7067-265X}{ORCID})

Authors:
\itemize{
  \item Juan Carlos Baez \email{juancarlos.baez@bc3research.org} (\href{https://orcid.org/0000-0002-7067-265X}{ORCID})
  \item Jon Sampedro \email{jon.sampedro@bc3research.org} (\href{https://orcid.org/0000-0002-2277-1530}{ORCID})
  \item Guillermo Pardo \email{guillermo.pardo@bc3research.org} (\href{https://orcid.org/0000-0002-7961-8457}{ORCID})
  \item Agustín del Prado \email{agustin.delprado@bc3research.org} (\href{https://orcid.org/0000-0003-3895-4478}{ORCID})
}

}
\keyword{internal}
