% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_data.R
\name{herdr_get_forages_path}
\alias{herdr_get_forages_path}
\title{Locate the forages dataset (forages.parquet)}
\usage{
herdr_get_forages_path(data_dir = "user_data")
}
\arguments{
\item{data_dir}{Character. Path to the user data directory. Default is \code{"user_data"}.}
}
\value{
Absolute path to forages parquet file.
}
\description{
Checks in order:
\enumerate{
  \item Local data directory (\code{file.path(data_dir, "forages.parquet")})
  \item Package bundled extdata (\code{system.file("extdata", "forages.parquet", package = "herdr")})
}
}
