% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{herdr_init}
\alias{herdr_init}
\title{Initialize herdr project folders and example data}
\usage{
herdr_init()
}
\description{
This function creates the 'user_data', 'Examples', and 'output' folders in your
current working directory and populates them with the default files from the package.
}
