% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_dataset}
\alias{load_dataset}
\title{Load dataset from user_data or package default}
\usage{
load_dataset(name)
}
\arguments{
\item{name}{Dataset name without extension, e.g., "diet", "categories"}
}
\value{
Data frame
}
\description{
Load dataset from user_data or package default
}
