% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcf}
\alias{mcf}
\title{Factores de Conversión de Metano (MCF)}
\format{
Coeficientes MCF según el tipo de gestión.
}
\usage{
mcf
}
\description{
Factores de Conversión de Metano (MCF)
}
\keyword{datasets}
