% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n2o_direct}
\alias{n2o_direct}
\title{Emisiones de N2O directas}
\format{
Resultados o factores para óxido nitroso directo.
}
\usage{
n2o_direct
}
\description{
Emisiones de N2O directas
}
\keyword{datasets}
