% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n2o_indirect}
\alias{n2o_indirect}
\title{Emisiones de N2O indirectas}
\format{
Resultados o factores para óxido nitroso indirecto.
}
\usage{
n2o_indirect
}
\description{
Emisiones de N2O indirectas
}
\keyword{datasets}
