% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rate_parameters}
\alias{rate_parameters}
\title{Parámetros de tasa (Rate parameters)}
\format{
Constantes de tasa utilizadas en las ecuaciones.
}
\usage{
rate_parameters
}
\description{
Parámetros de tasa (Rate parameters)
}
\keyword{datasets}
