% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_herdr_app}
\alias{run_herdr_app}
\title{Launch the herdr Interactive Application}
\usage{
run_herdr_app()
}
\description{
Opens the local herdr graphical user interface (GUI) in your default web browser.
This allows you to run impact assessments offline keeping your data private.
}
