% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_emissions_manure_CH4.R
\name{calculate_CH4_manure}
\alias{calculate_CH4_manure}
\title{Calculate CH₄ Emissions from Manure Management (Refactored)}
\usage{
calculate_CH4_manure(saveoutput = TRUE)
}
\description{
Computes CH₄ emissions from manure based on Volatile Solids (VS),
population, and management factors (B0, MCF, AWMS). (IPCC Eq 10.23).
}
