% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_calculate_DMI.R
\name{calculate_DMI}
\alias{calculate_DMI}
\title{Calculate Dry Matter Intake (DMI)}
\usage{
calculate_DMI(saveoutput = TRUE)
}
\arguments{
\item{saveoutput}{If TRUE (default) the results are saved in the output folder.}
}
\description{
Computes daily Dry Matter Intake (kg DM/day) based on metabolic demand (GE/ED).
}
