% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_calculate_GE.R
\name{calculate_ge}
\alias{calculate_ge}
\title{Calculate Gross Energy (GE)}
\usage{
calculate_ge(saveoutput = TRUE)
}
\arguments{
\item{saveoutput}{If TRUE (default) the results are saved in the output folder.}
}
\description{
Computes gross energy requirements by aggregating Net Energy (NE) components.
}
