% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_land_use.R
\name{calculate_land_use}
\alias{calculate_land_use}
\title{Calculate land use}
\usage{
calculate_land_use(
  automatic_cycle = FALSE,
  saveoutput = TRUE,
  farm_country = "Spain",
  year = 2022
)
}
\arguments{
\item{automatic_cycle}{Logical. If TRUE, uses the built-in model for automatic farm cycle calculation. Default is FALSE.}

\item{saveoutput}{If TRUE (default) the results are saved in the output folder.}

\item{farm_country}{Character. The country of the farm/study (e.g., "Spain"). Default is "Spain".}

\item{year}{Numeric. The reference year for FAO trade data calculation if origins are missing. Default is 2022.}
}
\description{
Computes total land use (m2) per animal based on validated DMI and specific origin countries.
}
