% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_land_use.R
\name{calculate_land_use}
\alias{calculate_land_use}
\title{Calcular uso de tierra}
\usage{
calculate_land_use(animal = NULL, type = NULL, zone = NULL, saveoutput = TRUE)
}
\arguments{
\item{animal}{Character string (opcional). Tipo de ganado (columna `animal_type`). Default NULL.}

\item{type}{Character string (opcional). Subtipo de ganado (columna `animal_subtype`). Default NULL.}

\item{zone}{Character vector (opcional). Filtrar por zona (columna `Zone`). Default NULL.}

\item{saveoutput}{Logical. Si TRUE, guarda el resultado en "output/Land_use.csv". Default FALSE.}
}
\value{
Tibble con uso de tierra por ingrediente.
}
\description{
Calcula el uso de tierra (m²) basado en consumo de ingredientes y rendimiento de cultivos.
}
\examples{
\donttest{
calculate_land_use(animal = "Cattle", type = "Dairy", zone = c("A"), saveoutput = TRUE)
}
}
