% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_land_use.R
\name{calculate_land_use}
\alias{calculate_land_use}
\title{Calculate land use}
\usage{
calculate_land_use(
  animal = NULL,
  type = NULL,
  zone = NULL,
  saveoutput = TRUE,
  population_df = NULL
)
}
\arguments{
\item{animal}{Character string/vector (optional). Animal type (column `animal_type`). Default NULL.}

\item{type}{Character string/vector (optional). Subtype of animal (column `animal_subtype`). Default NULL.}

\item{zone}{Character vector (optional). Filter by zone (column `zone`). Default NULL.}

\item{saveoutput}{Logical. If TRUE, saves the result to "output/Land_use.csv". Default TRUE.}

\item{population_df}{(Opcional) Un dataframe de población ya calculado.
Si es NULL (por defecto), la función llamará a `calculate_population()` interactivamente.}
}
\value{
Tibble with total land use per ingredient.
}
\description{
Computes total land use (m²) based on ingredient consumption, crop yields,
and animal population.
}
