% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_N2O_indirect_leaching.R
\name{calculate_N2O_indirect_leaching}
\alias{calculate_N2O_indirect_leaching}
\title{Calculate indirect N2O emissions from leaching}
\usage{
calculate_N2O_indirect_leaching(automatic_cycle = FALSE, saveoutput = TRUE)
}
\arguments{
\item{automatic_cycle}{Logical. If TRUE, uses the built-in model for automatic farm cycle calculation. Default is FALSE.}

\item{saveoutput}{If TRUE (default) the results are saved in the output folder.}
}
\description{
Computes indirect N2O emissions derived from nitrogen leaching (IPCC Eq 10.27 and 10.29).
}
