% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_N2O_indirect_volatilization.R
\name{calculate_N2O_indirect_volatilization}
\alias{calculate_N2O_indirect_volatilization}
\title{Calculate indirect N₂O emissions from volatilization (Refactored)}
\usage{
calculate_N2O_indirect_volatilization(saveoutput = TRUE)
}
\description{
Computes indirect N₂O emissions derived from volatilization of excreted nitrogen.
}
