% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_energy_net.R
\name{calculate_NE_pregnancy}
\alias{calculate_NE_pregnancy}
\title{Calculate Net Energy for Pregnancy (NE_pregnancy)}
\usage{
calculate_NE_pregnancy(saveoutput = TRUE)
}
\description{
Calculate Net Energy for Pregnancy (NE_pregnancy)
}
