% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_energy_requirements.R
\name{calculate_NE_pregnancy}
\alias{calculate_NE_pregnancy}
\title{Calculate Net Energy for Pregnancy (NE_pregnancy)}
\usage{
calculate_NE_pregnancy(animal = NULL, type = NULL, saveoutput = TRUE)
}
\arguments{
\item{animal}{Character (optional)}

\item{type}{Character (optional, only for Cattle)}

\item{saveoutput}{Logical (optional). If TRUE, saves the result as CSV. Default FALSE.}
}
\value{
Tibble with code, animal_type, animal_subtype, c_pregnancy, NEm, NE_pregnancy
}
\description{
Computes NE_pregnancy with species-specific rules.
}
