% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_energy_net.R
\name{calculate_NE_work}
\alias{calculate_NE_work}
\title{Calculate Net Working Energy (NE_work)}
\usage{
calculate_NE_work(saveoutput = TRUE)
}
\arguments{
\item{saveoutput}{Logical (optional). If TRUE, saves the result as CSV. Default TRUE.}
}
\value{
A tibble with NE_work for all relevant animal categories.
}
\description{
NE_work = NEm * hours of activity.
}
