% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_energy_net.R
\name{calculate_NE_work}
\alias{calculate_NE_work}
\title{Calculate Net Energy for Work (NE_work)}
\usage{
calculate_NE_work(saveoutput = TRUE)
}
\arguments{
\item{saveoutput}{If TRUE (default) the results are saved in the output folder.}
}
\description{
Computes NE_work based on NEm and hours of activity.
}
