% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_energy_net.R
\name{calculate_NEg}
\alias{calculate_NEg}
\title{Calculate Net Energy for Growth (NEg)}
\usage{
calculate_NEg(saveoutput = TRUE)
}
\arguments{
\item{saveoutput}{Logical (optional). If TRUE, saves the result as CSV. Default TRUE.}
}
\value{
A tibble with NEg for all animal categories.
}
\description{
Computes NEg for all animals.
}
