% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_energy_requirements.R
\name{calculate_NEl}
\alias{calculate_NEl}
\title{Calculate Net Energy for Lactation (NEl)}
\usage{
calculate_NEl(animal = NULL, type = NULL, saveoutput = TRUE)
}
\arguments{
\item{animal}{Character string (optional). Animal type.}

\item{type}{Character string (optional). Only for Cattle subtype.}

\item{saveoutput}{Logical (optional). If TRUE, saves the result as CSV. Default FALSE.}
}
\value{
A tibble with code, animal_type, animal_subtype, Milk_yield_kg_day_head, fat_content, NEl
}
\description{
Computes NEl based on milk yield and fat content.
}
