% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_energy_net.R
\name{calculate_NEm}
\alias{calculate_NEm}
\title{Calculate Net Energy for Maintenance (NEm)}
\usage{
calculate_NEm(saveoutput = TRUE)
}
\arguments{
\item{saveoutput}{Logical. Save CSV? Default TRUE}
}
\value{
Tibble with NEm for all animal categories.
}
\description{
Computes NEm for all livestock animals based on average weight and
CFI coefficients found in the input files.
}
