% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_cattle.R
\name{calculate_population_cattle}
\alias{calculate_population_cattle}
\title{Calculates the full cattle population structure (Internal Helper)}
\usage{
calculate_population_cattle(census_cattle, rate_parameters, categories)
}
\description{
Calculates the full cattle population structure (Internal Helper)
}
