% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_cattle.R
\name{calculate_population_cattle}
\alias{calculate_population_cattle}
\title{Calculates the full cattle population structure (Internal Helper)}
\usage{
calculate_population_cattle(census_cattle, rate_parameters, categories)
}
\arguments{
\item{census_cattle}{Edit in csv}

\item{rate_parameters}{Edit in csv}

\item{categories}{Edit in csv}
}
\description{
Calculates the full cattle population structure (Internal Helper)
}
