% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_goat.R
\name{calculate_population_goat}
\alias{calculate_population_goat}
\title{Calculates the full goat population structure (Internal Helper)}
\usage{
calculate_population_goat(census_goat, rate_parameters)
}
\arguments{
\item{census_goat}{Edit in csv}

\item{rate_parameters}{Edit in csv}
}
\description{
Calculates the full goat population structure (Internal Helper)
}
