% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_sheep.R
\name{calculate_population_sheep}
\alias{calculate_population_sheep}
\title{Calculates the full sheep population structure (Internal Helper)}
\usage{
calculate_population_sheep(census_sheep, rate_parameters)
}
\description{
Calculates the full sheep population structure (Internal Helper)
}
