% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_user_data.R
\name{download_templates}
\alias{download_templates}
\title{Descarga los archivos de template desde GitHub}
\usage{
download_templates(dest_folder = "user_data")
}
\arguments{
\item{dest_folder}{El nombre de la carpeta donde se guardarán los
templates. Por defecto es "user_data".}
}
\value{
El path absoluto a la carpeta `dest_folder` donde se
  guardaron los archivos.
}
\description{
Descarga los archivos de ejemplo (census, categories, rate_parameters)
desde el repositorio de GitHub y los guarda en una carpeta local.
}
\examples{
\dontrun{
  # Llama a la función para descargar los archivos
  ruta_datos <- download_templates()

  # Ver los archivos descargados
  list.files(ruta_datos)

  # Leer uno de los archivos
  read.csv(file.path(ruta_datos, "census_c_s.csv"))
}

}
