% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_summary.R
\name{generate_impact_assessment}
\alias{generate_impact_assessment}
\title{Summarize CH4, N2O Emissions, and Land Use}
\usage{
generate_impact_assessment(
  automatic_cycle = FALSE,
  region = NULL,
  subregion = NULL,
  animal = NULL,
  type = NULL,
  class_flex = NULL,
  saveoutput = TRUE,
  group_by_identification = TRUE,
  farm_country = "Spain",
  year = 2024,
  data_dir = "user_data"
)
}
\arguments{
\item{automatic_cycle}{Logical. TRUE for built-in model, FALSE for manual livestock_census.csv.}

\item{region}{Character/Numeric vector to filter.}

\item{subregion}{Character vector to filter.}

\item{animal}{Livestock type (animal_type).}

\item{type}{Livestock subtype (animal_subtype).}

\item{class_flex}{Management class (e.g., 'grazing', 'stall').}

\item{saveoutput}{If TRUE saves to output folder.}

\item{group_by_identification}{If TRUE returns by animal_tag.}

\item{farm_country}{Character. The country of the farm/study (e.g., "Spain"). Default is "Spain".}

\item{year}{Numeric. The reference year for FAO trade data calculation if origins are missing. Default is 2022.}

\item{data_dir}{Path to the directory containing input CSV/data files. Defaults to `"user_data"`.}
}
\description{
Summarize CH4, N2O Emissions, and Land Use
}
