% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_summary.R
\name{generate_impact_assessment}
\alias{generate_impact_assessment}
\title{Summarize CH4, N2O Emissions, and Land Use}
\usage{
generate_impact_assessment(
  automatic_cycle = FALSE,
  region = NULL,
  subregion = NULL,
  animal = NULL,
  type = NULL,
  class_flex = NULL,
  saveoutput = TRUE,
  group_by_identification = TRUE
)
}
\arguments{
\item{automatic_cycle}{Logical. TRUE for built-in model, FALSE for manual livestock_census.csv.}

\item{region}{Character/Numeric vector to filter.}

\item{subregion}{Character vector to filter.}

\item{animal}{Livestock type (animal_type).}

\item{type}{Livestock subtype (animal_subtype).}

\item{class_flex}{Management class (e.g., 'grazing', 'stall').}

\item{saveoutput}{If TRUE saves to output folder.}

\item{group_by_identification}{If TRUE returns by animal_tag.}
}
\description{
Summarize CH4, N2O Emissions, and Land Use
}
