% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_summary.R
\name{generate_impact_assessment}
\alias{generate_impact_assessment}
\title{Summarize CH₄, N₂O Emissions, and Land Use}
\usage{
generate_impact_assessment(
  group = NULL,
  zone = NULL,
  animal = NULL,
  type = NULL,
  saveoutput = TRUE,
  group_by_identification = TRUE
)
}
\arguments{
\item{group}{Character/Numeric vector (optional). Groups to filter. Default `NULL`.}

\item{zone}{Character vector (optional). Zones to filter. Default `NULL`.}

\item{animal}{Character string (optional). Livestock type (`animal_type`). Default `NULL`.}

\item{type}{Character string (optional). Livestock subtype (`animal_subtype`). Default `NULL`.}

\item{saveoutput}{If TRUE (default) the results are saved in the output folder.}

\item{group_by_identification}{Logical. If `TRUE`, returns results at the `identification` level. If `FALSE`, aggregates all `identification`s.}
}
\value{
A `tibble` with columns for each emission type and land use.
}
\description{
Calculates a summary of CH₄ and N₂O emissions (direct and indirect) and land use,
grouped by code, animal type, subtype, and zone. Optionally allows for
aggregation to a more general level.
}
