% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animal_GEI_LU.R, R/herdr-package.R
\docType{package}
\name{herdr-package}
\alias{herdr-package}
\title{herdr: Package for Animal Greenhouse Gas Emissions and Land Use}
\description{
This package provides tools for calculating and analyzing greenhouse gas
emissions and land use metrics for animal systems.

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Description: Provides a comprehensive set of tools for calculating livestock greenhouse gas emissions based on Spanish national standards. The package includes functions to process animal census data for cattle, sheep, and goats, integrating specific emission factors and rate parameters. It facilitates the estimation of direct and indirect emissions, such as methane and nitrous oxide, following standardized methodologies for environmental reporting.
}
\details{
For more information, see:
\itemize{
\item GitHub: \url{https://github.com/JuanCBM99/herdr}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JuanCBM99/herdr}
}


Useful links:
\itemize{
  \item \url{https://github.com/JuanCBM99/herdr}
}

}
\author{
\strong{Maintainer}: Juan Carlos Baez \email{juancarlos.baez@bc3research.org} (\href{https://orcid.org/0000-0002-7067-265X}{ORCID})

}
\keyword{internal}
