% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_herdr_results}
\alias{plot_herdr_results}
\title{Plot herdr results dynamically with custom grouping and premium aesthetics}
\usage{
plot_herdr_results(
  df,
  group_cols = c("animal_tag", "region", "subregion", "class_flex"),
  func_name = NULL
)
}
\arguments{
\item{df}{Dataframe containing the results.}

\item{group_cols}{Character vector of columns to group the plot by.}

\item{func_name}{Name of the function that generated the data.}
}
\value{
A ggplot2 object.
}
\description{
Plot herdr results dynamically with custom grouping and premium aesthetics
}
